services/video-generation/requirements-enhanced.txt
# Enhanced Requirements for Video Generation Service
# Optional dependencies for TOP AI technologies 2025
# Install with: pip install -r requirements-enhanced.txt

# ========================================
# BASE REQUIREMENTS (from requirements.txt)
# ========================================
-r requirements.txt

# ========================================
# PYTORCH LIGHTNING (10x faster training)
# ========================================
pytorch-lightning==2.1.3
lightning[extra]==2.1.3
torchmetrics==1.2.1

# ========================================
# ONNX RUNTIME (cross-platform optimization)
# ========================================
onnx==1.15.0
onnxruntime-gpu==1.17.0
onnxconverter-common==1.14.0

# ========================================
# VIDEO-SPECIFIC OPTIMIZATION
# ========================================
# FFmpeg bindings
ffmpeg-python==0.2.0

# Video processing acceleration
av==11.0.0

# ========================================
# OPTIMIZATION & ACCELERATION
# ========================================
# Mixed precision training
optimum==1.16.1
bitsandbytes==0.41.3

# ========================================
# CONFIGURATION & UTILITIES
# ========================================
pyyaml==6.0.1  # For feature flags
hydra-core==1.3.2  # Advanced configuration management

# ========================================
# MONITORING & LOGGING
# ========================================
wandb==0.16.2  # Weights & Biases integration (optional)
tensorboard==2.15.1
